Quantities

This section describes the Quantities submodule of Alicorn. The module is concerned with defining and manipulating physical quantities.

Unless stated otherwise, all types, functions, and constants defined in the submodule are exported by Alicorn.

Contents

Quantities

Unit conversion

Unit conversion

InternalUnits

Quantities

Alicorn.Quantities.AbstractQuantity — Type

AbstractQuantity{T}

Abstract supertype for all types that represent a scalar physical quantity.

Alicorn.Quantities.SimpleQuantity — Type

SimpleQuantity{T<:Number} <: AbstractQuantity{T}

A physical quantity consisting of a scalar value and a physical unit.

The value field of a SimpleQuantity{T} is of type T, which needs to be a subtype of Number.

Fields

value::T: value of the quantity

unit::Unit: unit of the quantity

Constructors

SimpleQuantity(value::T, abstractUnit::AbstractUnit) where T <: Number
SimpleQuantity(value::T) where T <: Number
SimpleQuantity(abstractUnit::AbstractUnit)
SimpleQuantity(simpleQuantity::SimpleQuantity)

If no AbstractUnit is passed to the constructor, the Alicorn.unitlessUnit is used by default. If no value is passed to the constructor, the value is set to 1 by default.

Examples

The quantity $7\,\mathrm{nm}$ (seven nanometers) can be constructed using the constructor method as follows:

julia> ucat = UnitCatalogue() ;

julia> nanometer = ucat.nano * ucat.meter
UnitFactor nm

julia> quantity = SimpleQuantity(7, nanometer)
7 nm

Alternatively, $7\,\mathrm{nm}$ can be constructed arithmetically:

julia> ucat = UnitCatalogue() ;

julia> nm = ucat.nano * ucat.meter
UnitFactor nm

julia> quantity = 7nm
7 nm

Alicorn.Quantities.Quantity — Type

Quantity{T<:Number} <: AbstractQuantity{T}

A physical quantity consisting of a number, a Dimension object representing the physical dimension, and an InternalUnits object representing the units with respect to which the seven basic dimensions of the SI system are measured.

The value field of a Quantity{T} is of type T, which needs to be a subtype of Number.

Fields

value::T: value of the quantity

dimension::Dimension: physical dimension of the quantity

internalUnits::InternalUnits: set of units with respect to which the seven

basic dimensions of the SI system are measured.

Constructors

Quantity(value::T, dimension::Dimension, internalUnits::InternalUnits) where T <: Number
Quantity(simpleQuantity::SimpleQuantity, internalUnits::InternalUnits)

Base.:* — Method

Base.:*(value::Number, abstractUnit::AbstractUnit)

Combine value and abstractUnit to form a physical quantity of type SimpleQuantity.

Example

julia> ucat = UnitCatalogue() ;

julia> 3.5 * ucat.tesla
3.5 T

Base.:/ — Method

Base.:/(value::Number, abstractUnit::AbstractUnit)

Combine value and abstractUnit to form a physical quantity of type SimpleQuantity.

Example

julia> ucat = UnitCatalogue() ;

julia> 3.5 / ucat.second
3.5 s^-1

Unit conversion

Alicorn.Quantities.inUnitsOf — Method

inUnitsOf(quantity::AbstractQuantity, unit::AbstractUnit)::SimpleQuantity

Express quantity as an object of type SimpleQuantity in terms of the unit specified by unit.

Raises Exceptions

Alicorn.Exceptions.DimensionMismatchError: if the dimensions of quantity and unit do not agree

Alicorn.Quantities.inBasicSIUnits — Method

inBasicSIUnits(quantity::AbstractQuantity)::SimpleQuantity

Express quantity as an object of type SimpleQuantity using the seven basic SI units.

Alicorn.Quantities.valueInUnitsOf — Method

valueInUnitsOf(quantity::AbstractQuantity, unit::AbstractUnit)

Returns the numerical value of quantity expressed in units of unit.

Raises Exceptions

Alicorn.Exceptions.DimensionMismatchError: if the dimensions of quantity and unit do not agree

Alicorn.Quantities.valueInUnitsOf — Method

valueInUnitsOf(quantity::AbstractQuantity, simpleQuantity::SimpleQuantity)

Returns the numerical value of quantity expressed in units of simpleQuantity.

The result is equivalent to valueOfDimensionless(quantity / simpleQuantity).

Raises Exceptions

Alicorn.Exceptions.DimensionMismatchError: if the dimensions of quantity and simpleQuantity do not agree

Alicorn.Quantities.valueOfDimensionless — Method

valueOfDimensionless(quantity::AbstractQuantity)

Strips the unit from a dimensionless quantity and returns its bare value.

Raises Exceptions

Alicorn.Exceptions.DimensionMismatchError: if quantity is not dimensionless

Base.:* — Method

Base.:*(quantity::AbstractQuantity, unit::AbstractUnit)
Base.:*(unit::AbstractUnit, quantity::AbstractQuantity)

Modify the unit of quantity by multiplying it with unit.

Base.:/ — Method

Base.:/(quantity::AbstractQuantity, unit::AbstractUnit)
Base.:/(unit::AbstractUnit, quantity::AbstractQuantity)

Modify the unit of quantity by dividing it by unit, or vice versa.

Alicorn.Quantities.valueOfDimensionless — Method

valueOfDimensionless(quantity::AbstractQuantity)

Strips the unit from a dimensionless quantity and returns its bare value.

Raises Exceptions

Alicorn.Exceptions.DimensionMismatchError: if quantity is not dimensionless

Base.:== — Method

Base.:(==)(simpleQuantity1::SimpleQuantity, simpleQuantity2::SimpleQuantity)

Returns true if simpleQuantity1 and simpleQuantity2 are of equal value.

If necessary, simpleQuantity2 is expressed in units of simpleQuantity1.unit before the comparison. Note that the conversion oftens lead to rounding errors that render simpleQuantity1 not equal simpleQuantity2.

Raises Exceptions

Alicorn.Exceptions.DimensionMismatchError: if simpleQuantity1 and

simpleQuantity2 are not of the same dimension

Examples

julia> ucat = UnitCatalogue() ;

julia> q1 = 7 * ucat.meter
7 m

julia> q2 = 700 * (ucat.centi * ucat.meter)
700 cm

julia> q1 == q1
true

julia> q1 == q2
true

Base.:== — Method

Base.:(==)(quantity1::Quantity, quantity2::Quantity)

Compare two Quantity objects.

The two quantities are equal if their values, their dimensions, and their internal units are equal. Note that the units are not converted during the comparison.

Base.:+ — Method

Base.:+(simpleQuantity1::SimpleQuantity, simpleQuantity2::SimpleQuantity)

Add two SimpleQuantities.

The resulting quantity is expressed in units of simpleQuantity1.

Raises Exceptions

Alicorn.Exceptions.DimensionMismatchError: if simpleQuantity1 and simpleQuantity2 are of different dimensions

Base.:- — Method

Base.:-(simpleQuantity1::SimpleQuantity, simpleQuantity2::SimpleQuantity)

Subtract two SimpleQuantities.

The resulting quantity is expressed in units of simpleQuantity1.

Raises Exceptions

Alicorn.Exceptions.DimensionMismatchError: if simpleQuantity1 and simpleQuantity2 are of different dimensions

Base.isless — Method

Base.isless(simpleQuantity1::SimpleQuantity, simpleQuantity2::SimpleQuantity)

Returns true if simpleQuantity1 is of lesser value than simpleQuantity2.

If necessary, simpleQuantity2 is expressed in units of simpleQuantity1.unit before the comparison. Note that the conversion often leads to rounding errors.

Raises Exceptions

Alicorn.Exceptions.DimensionMismatchError: if simpleQuantity1 and

simpleQuantity2 are not of the same dimension

Base.isapprox — Method

Base.isapprox(simpleQuantity1::SimpleQuantity, simpleQuantity2::SimpleQuantity; rtol::Real = sqrt(eps()) )

Returns isapprox(simpleQuantity1.value, simpleQuantity2.value, rtol=rtol).

If necessary, simpleQuantity2 is expressed in units of simpleQuantity1.unit before the comparison. Note that the conversion often leads to rounding errors.

Raises Exceptions

Alicorn.Exceptions.DimensionMismatchError: if simpleQuantity1 and

simpleQuantity2 are not of the same dimension

Base.zero — Method

Base.zero(numberType::Type, unit::AbstractUnit)

Return a SimpleQuantity with unit unit and value zero(numberType).

Raises Exceptions

Alicorn.DomainError: if numberType is not a subtype of Number.

Quantity Arrays

Alicorn.Quantities.AbstractQuantityArray — Type

AbstractQuantity{T,N} <: AbstractArray{T,N}

Abstract supertype for all types that represent a physical quantity array.

Alicorn.Quantities.SimpleQuantityArray — Type

SimpleQuantityArray{T<:Number,N} <: AbstractQuantityArray{T,N}

A physical quantity consisting of a number array and a physical unit.

The value field of a SimpleQuantityArray{T,N} is of type Array{T,N}. T needs to be a subtype of Number.

Fields

value::Array{T,N}: value of the quantity

unit::Unit: unit of the quantity

Constructors

``` SimpleQuantityArray(value::AbstractArray{T,N}, abstractUnit::AbstractUnit) where {T<:Number, N} SimpleQuantityArray(value::AbstractArray{T,N}) where {T<:Number, N} SimpleQuantityArray(simpleQuantity::SimpleQuantity) SimpleQuantityArray(sqArray::SimpleQuantityArray)

Alicorn.Quantities.QuantityArray — Type

QuantityArray{T<:Number,N} <: AbstractQuantityArray{T,N}

A physical quantity consisting of an array, a Dimension object representing the physical dimension, and an InternalUnits object representing the units with respect to which the seven basic dimensions of the SI system are measured.

The value field of a QuantityArray{T,N} is of type Array{T,N}. T needs to be a subtype of Number.

Fields

value::Array{T,N}: value of the quantity

dimension::Dimension: physical dimension of the quantity

internalUnits::InternalUnits: set of units with respect to which the seven

basic dimensions of the SI system are measured.

Constructors

``` QuantityArray(value::AbstractArray{T,N}, dimension::Dimension, internalUnits::InternalUnits) QuantityArray(sqArray::SimpleQuantityArray, internalUnits::InternalUnits)

Base.:* — Method

Base.:*(value::AbstractArray{T,N}, abstractUnit::AbstractUnit) where {T<:Number, N}

Combine the array value and abstractUnit to form a physical quantity of type SimpleQuantityArray.

Example

julia> ucat = UnitCatalogue() ;

julia> [3.5, 4.6] * ucat.tesla
2-element SimpleQuantityVector{Float64} of unit T:
 3.5
 4.6

Base.:/ — Method

Base.:/(value::AbstractArray{T,N}, abstractUnit::AbstractUnit) where {T<:Number, N}

Combine the array value and abstractUnit to form a physical quantity of type SimpleQuantityArray.

Example

julia> ucat = UnitCatalogue() ;

julia> [3.5, 4.6] / ucat.second
2-element SimpleQuantityVector{Float64} of unit s^-1:
 3.5
 4.6

Unit conversion

Alicorn.Quantities.inUnitsOf — Method

inUnitsOf(qArray::AbstractQuantityArray, unit::AbstractUnit)::SimpleQuantityArray

Express qArray as an object of type SimpleQuantityArray in terms of the unit specified by unit.

Raises Exceptions

Alicorn.Exceptions.DimensionMismatchError: if the dimensions of qArray and unit do not agree

Alicorn.Quantities.inBasicSIUnits — Method

inBasicSIUnits(qArray::AbstractQuantity)::SimpleQuantityArray

Express qArray as an object of type SimpleQuantityArray using the seven basic SI units.

Alicorn.Quantities.valueInUnitsOf — Method

valueInUnitsOf(qArray::AbstractQuantityArray, unit::AbstractUnit)

Returns the numerical value of qArray expressed in units of unit.

Raises Exceptions

Alicorn.Exceptions.DimensionMismatchError: if the dimensions of quantity and unit do not agree

Alicorn.Quantities.valueInUnitsOf — Method

valueInUnitsOf(quantityArray::AbstractQuantityArray, simpleQuantity::SimpleQuantity)

Returns the numerical value of quantityArray expressed in units of simpleQuantity.

The result is equivalent to valueOfDimensionless(quantityArray / simpleQuantity).

Raises Exceptions

Alicorn.Exceptions.DimensionMismatchError: if the dimensions of quantityArray and simpleQuantity do not agree

Alicorn.Quantities.valueOfDimensionless — Method

valueOfDimensionless(qArray::AbstractQuantityArray)

Strips the unit from a dimensionless quantity array and returns its bare value.

Raises Exceptions

Alicorn.Exceptions.DimensionMismatchError: if qArray is not dimensionless

Base.:* — Method

Base.:*(qArray::AbstractQuantityArray, unit::AbstractUnit)
Base.:*(unit::AbstractUnit, qArray::AbstractQuantityArray)

Modify the unit of qArray by multiplying it with unit.

Base.:/ — Method

Base.:/(qArray::AbstractQuantityArray, unit::AbstractUnit)
Base.:/(unit::AbstractUnit, qArray::AbstractQuantityArray)

Modify the unit of qArray by dividing it by unit, or vice versa.

Base.:== — Method

Base.:(==)(sqArray1::SimpleQuantityArray, sqArray2::SimpleQuantityArray)

Returns true if sqArray1 and sqArray2 are of equal value and dimension.

If necessary, sqArray2 is expressed in units of sqArray1.unit before the comparison. Note that the conversion oftens lead to rounding errors that render sqArray1 not equal sqArray2.

Raises Exceptions

Alicorn.Exceptions.DimensionMismatchError: if sqArray1 and

sqArray2 are not of the same dimension ```

Base.:== — Method

Base.:(==)(qArray1::QuantityArray, qArray2::QuantityArray)

Compare two QuantityArray objects.

The two quantities are equal if their values, their dimensions, and their internal units are equal. Note that the units are not converted during the comparison.

Base.isapprox — Method

Base.isapprox(sqArray1::SimpleQuantityArray, sqArray2::SimpleQuantityArray; rtol::Real = sqrt(eps()) )

Returns isapprox(sqArray1.value, sqArray2.value, rtol=rtol).

If necessary, sqArray2 is expressed in units of sqArray1.unit before the comparison. Note that the conversion often leads to rounding errors.

Raises Exceptions

Alicorn.Exceptions.DimensionMismatchError: if sqArray1 and

sqArray2 are not of the same dimension

Base.:+ — Method

Base.:+(sqArray1::SimpleQuantityArray, sqArray2::SimpleQuantityArray)

Add two SimpleQuantityArrays.

The resulting quantity is expressed in units of sqArray1.

Raises Exceptions

Alicorn.Exceptions.DimensionMismatchError: if sqArray1 and sqArray2 are of different dimensions

Base.:- — Method

Base.:-(sqArray1::SimpleQuantityArray, sqArray2::SimpleQuantityArray)

Subtract two SimpleQuantityArrays.

The resulting quantity is expressed in units of sqArray1.

Raises Exceptions

Alicorn.Exceptions.DimensionMismatchError: if sqArray1 and sqArray2 are of different dimensions

InternalUnits

Alicorn.Quantities.InternalUnits — Type

InternalUnits

A set of seven SimpleQuantity objects which represent a choice of units with respect to which the seven basic physical dimensions of the SI system are measured.

Fields

mass::SimpleQuantity: unit of mass

length::SimpleQuantity: unit of length

time::SimpleQuantity: unit of time

current::SimpleQuantity: unit of current

temperature::SimpleQuantity: unit of temperature

amount::SimpleQuantity: unit of amount

luminousIntensity::SimpleQuantity: unit of luminousIntensity

Constructors

InternalUnits(::SimpleQuantity, ::SimpleQuantity, ::SimpleQuantity, ::SimpleQuantity, ::SimpleQuantity, ::SimpleQuantity, ::SimpleQuantity)
InternalUnits(; mass = 1*kilogram, length = 1*meter, time = 1*second, current = 1*ampere,temperature = 1*kelvin, amount = 1*mol, luminousIntensity = 1*candela)

# Raises Exceptions
- `Core.DomainError`: if attempting to initialize any field with a quantity
of a value that is zero, infinite, or not real
- `Exceptions.DimensionMismatchError`: if attempting to initialize any field
with a quantity whose dimension does not match the physical dimension the field
represents

# Examples
The following `InternalUnits` measure lengths in units of ``3 cm`` and uses
the basic SI units for all other dimensions:

jldoctest julia> ucat = UnitCatalogue(); cm = ucat.centi * ucat.meter ;

julia> InternalUnits(length = 3cm) InternalUnits mass unit: 1 kg length unit: 3 cm time unit: 1 s current unit: 1 A temperature unit: 1 K amount unit: 1 mol luminous intensity unit: 1 cd ```

Base.:== — Method

Base.:(==)(internalUnits1::InternalUnits, internalUnits2::InternalUnits)

Compare two InternalUnits objects.

Two InternalUnits objects are equal if the SimpleQuantity objects representing the unit for each of the seven physical dimensions are equal.

Alicorn.Quantities.internalUnitForDimension — Method

internalUnitForDimension(dimension::Dimension, internalUnits::InternalUnits)

Returns a SimpleQuantity representing the unit in which quantities of physical dimension dimension are measured according to internalUnits.